feat: Update Tour management with new date fields and refactor related components

+++ src/main/java/com/example/booking_tour/repository/TourRepository.java
@@ -3,6 +3,6 @@ package com.example.booking_tour.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.booking_tour.entity.Tour;
 
-public interface TourRepository extends JpaRepository<Tour, Long> {
+public interface TourRepository extends JpaRepository<Tour, Integer> {
     
 }
